app/requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
apispec==6.6.1
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.8.2
bs4==0.0.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.8
dateparser==1.2.0
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
dominate==2.9.1
email_validator==2.1.1
exceptiongroup==1.2.1
Flask==2.3.3
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-PageDown==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
html2docx==1.6.0
htmldocx==0.0.6
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.0.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
limits==3.12.0
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==0.28.2
mdurl==0.1.2
moment==0.12.1
openai==1.35.2
ordered-set==4.1.0
packaging==24.1
prison==0.2.1
prompt_toolkit==3.0.47
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.52
SQLAlchemy-Utils==0.41.2
times==0.7
tinycss2==1.3.0
tqdm==4.66.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vine==5.1.0
visitor==0.1.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
zibal==1.1.0
zipp==3.19.2
zibal
persian-tools
python-magic==0.4.27
PyMuPDF==1.24.7